Products are entirely dispatched; consumers' request is only used to test if a product should be sent. Added console and node transparency, more...
A consumer's request is not used to filter the incoming products any
more. Filter and FilterOut nodes can be used for this purpose.

A transparent node always requests its aggregate demand, which is the
union of its subscribed consumers' requests.

In the pre and post pipelines, if a pre/post does not require a
certain product, then the product is passed to the next pre/post. For
this reason a FilterOut should always request the products it want to
filter.